feat: API demo (product, customer, user)

+++ src/main/java/com/pos/hyper/repository/ProductRepository.java
@@ -2,7 +2,6 @@ package com.pos.hyper.repository;
 
 import com.pos.hyper.model.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.relational.core.sql.In;
 
-public interface ProductRepository extends JpaRepository<Product, Integer> {
+public interface ProductRepository extends JpaRepository<Product, Long> {
 }
